pyproject.toml
[tool.poetry]
name = "boids"
version = "0.1.0"
description = "A Python project for simulating and animating boids"
authors = ["Wojciech Kosnik-Kowalczuk <wojciech.kosnik.kowalczuk@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pygame = "^2.6.1"
numpy = "^2.1.3"
loading_bar = {git = "https://github.com/Awiomanik/UtilityScripts", subdirectory = "Python_scripts/loading_bar"}
pillow = "^11.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
